pyproject.toml
[project]
name = "price-watch"
dynamic = ["version"]
description = "商品価格を監視して通知するスクリプトです。"
authors = [
    { name = "KIMATA Tetsuya", email = "kimata@green-rabbit.net" }
]

readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "pyyaml>=6.0.3",
    "ruamel.yaml>=0.18.6",
    "my-lib @ git+https://github.com/kimata/my-py-lib@33f2bfc39b950ad30bda0525f4824aa316d8ece5",
    "selenium>=4.39.0",
    "undetected-chromedriver>=3.5.5",
    "setuptools>=75.0.0",
    "pillow>=12.1.0",
    "amazon-paapi5>=1.1.2",
    "pydub>=0.25.1",
    "speechrecognition>=3.14.5",
    "flask>=3.1.0",
    "flask-cors>=6.0.0",
    "flask-pydantic>=0.12.0",
    "docopt>=0.6.2",
    "pydantic>=2.10.0",
    "matplotlib>=3.8.0",
    "argon2-cffi>=23.1.0",
]

[project.scripts]
price-watch = "price_watch.cli.app:main"
price-watch-healthz = "price_watch.cli.healthz:main"
price-watch-webui = "price_watch.cli.webui:main"

[build-system]
requires = ["hatchling", "uv-dynamic-versioning"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "flaky>=3.8.1",
    "playwright>=1.57.0",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-html>=4.1.1",
    "pytest-mock>=3.15.1",
    "pytest-playwright>=0.7.2",
    "pytest-xdist>=3.8.0",
    "pytest-timeout>=2.4.0",
    "time-machine>=3.2.0",
    "mypy>=1.19.1",
    "pyright>=1.1.408",
    "ty>=0.0.1a7",
    "types-pytz>=2025.2.0",
    "types-paramiko>=4.0.0",
    "types-pyyaml>=6.0.12",
    "types-requests>=2.32.4",
    "pydantic",
]

[tool.hatch.build.targets.wheel]
packages = ["src/price_watch"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.version]
source = "uv-dynamic-versioning"

[tool.uv]
default-groups = ["dev"]

[tool.uv-dynamic-versioning]
vcs = "git"
style = "semver"
format = "{base}"

[tool.ruff]
line-length = 110

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP"]

[tool.pytest.ini_options]
minversion = "6.0"
# E2E テストはデフォルトで除外（外部サーバーが必要なため）
# E2E テスト実行: uv run pytest tests/e2e/ --host <host> --port <port>
addopts = "--verbose --timeout=300 --durations=10 --log-file-level=DEBUG --log-format=\"%(asctime)s %(levelname)s %(message)s\" --log-format=\"%(asctime)s %(levelname)s [%(filename)s:%(lineno)s %(funcName)s] %(message)s\" --capture=sys --junit-xml=reports/junit-report.xml --html=reports/pytest.html --self-contained-html --cov=src --cov-report=html --cov-report=term-missing --cov-report=xml:reports/coverage.xml --cov-report=lcov:reports/coverage.lcov --numprocesses=auto --ignore=tests/e2e"
testpaths = ["tests"]
filterwarnings = [
    "ignore:datetime\\.datetime\\.utcfromtimestamp\\(\\) is deprecated:DeprecationWarning",
    "ignore::DeprecationWarning:multiprocessing\\.popen_fork",
    # pydub の警告（Python 3.12+ でのエスケープシーケンスと audioop 非推奨）
    "ignore:invalid escape sequence:SyntaxWarning:pydub.*",
    "ignore:'audioop' is deprecated:DeprecationWarning:pydub.*",
    "ignore:unclosed database.*:ResourceWarning:rich.*",
    "ignore:unclosed database.*:ResourceWarning:influxdb_client.*",
    "ignore:unclosed database.*:ResourceWarning:coverage.*",
    "ignore:unclosed database.*:ResourceWarning:time_machine.*",
    "ignore:unclosed database.*:ResourceWarning:werkzeug.*",
    "ignore:unclosed database.*:ResourceWarning:_pytest.*",
    "ignore:unclosed database.*:ResourceWarning:typing.*",
    "ignore:unclosed database.*:ResourceWarning:multiprocessing.*",
    "ignore:unclosed database.*:ResourceWarning:logging.*",
    "ignore::ResourceWarning:_pytest.unraisableexception",
    "ignore::ResourceWarning:coverage.sqldata",
    "ignore::ResourceWarning:coverage.collector",
]

[tool.coverage.run]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.coverage.html]
directory = "reports/coverage"

[tool.mypy]
python_version = "3.12"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = false
exclude = [
    "^tests/",
    "^.venv/",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "amazon_paapi5.*",
    "selenium.*",
    "pydub.*",
    "speech_recognition.*",
]
ignore_missing_imports = true
follow_imports = "skip"

[tool.pyright]
venvPath = "."
venv = ".venv"
pythonVersion = "3.13"
include = ["src", "tests"]
reportMissingImports = false
reportArgumentType = false
reportOperatorIssue = false
reportIndexIssue = false
reportGeneralTypeIssues = false
reportAttributeAccessIssue = false

[tool.ty]
[tool.ty.rules]
